backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
attrs==23.1.0
autopep8==2.0.4
certifi==2024.7.4
cffi==1.15.1
channels==4.1.0
colorama==0.4.6
coverage==7.3.2
Cython==3.0.11
diff-match-patch==20230430
Django==4.2.1
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.3
django-import-export==4.1.1
django-nine==0.2.7
django-nose==1.4.7
django-silk==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.6.1
elastic-transport==8.15.0
elasticsearch==8.15.0
elasticsearch-dsl==8.15.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==20.1.0
freezegun==1.5.1
frozenlist==1.4.1
gprof2dot==2022.7.29
hypothesis==6.112.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
multidict==6.1.0
nose==1.3.7
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.5.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.9.0
pyrsistent==0.19.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
setuptools==70.0.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.20
websockets==13.1
yarl==1.13.1
